Refactor code && check is a video is a favorite video
